Fix a couple of memory stomper bugs, one of which was 64-bit only.

+++ Source/userinterface/cocoa/views/MTCompositedGLView.m
@@ -166,7 +166,7 @@
 - (void)drawRect:(NSRect)rect
 {
 	const NSRect* dirtyRects;
-	int dirtyRectCount;
+	NSInteger dirtyRectCount;
 
 	[self getRectsBeingDrawn:&dirtyRects count:&dirtyRectCount];
 	// punch a hole so that the OpenGL view shows through.
